Changed ClaimBlock to use metadata
Previously, ClaimBlocks were identified entirely on their material, meaning the material would have to be exclusively used for this role. After speaking with Connor, he confirmed that this may pose an issue.

Now, ClaimBlock ItemStacks are identified using 1.16's PersistentDataContainers, and Blocks are identified using metadata which is loaded on initialisation when the ClaimChunks are being loaded.

Related to issue #6 and #11.