Changed ClaimBlock duplicate test
Previously, to prevent duplicate ClaimBlocks from being given, it only checked against the material. This would mean that if a player had the a block of the same material (despite not being a claim block) in their inventory, it would be flagged as a duplicate. Now it checks the block's metadata.