Added random spawning
Players will now spawn (if they have no bed) to a random position which is not in a claimed chunk. The block will be one of the materials defined in NationVariables.goodSpawnBlocks. The block will also be the highest possible block at that X and Z. This means that you could theoretically spawn in a ravine with no ceiling if it has a goodSpawnBlock inside it. Though I don't see a simple way to fix that, and it should be very rare.