Added basic multi-language support
Added multi-language support for messages sent by the server to the player. This is currently done entirely through the player's locale as defined in their settings, though I would like to add a command to allow players to modify this.

I have not implemented multi-language support for subcommands yet, either, which I would like to add.

Related to issue #5.