Fixed nation map signed byte bug
Fixed a bug where Minecraft was being passed in negative co-ordinates for certain blocks. This is due to Java treating bytes as signed and I was saving the Y level of the chunk in a byte. Why they decided to do this is beyond my comprehension. Literally no other language does this and after some research it is widely regarded as a bad decision.

Updated POM version to v0.2.7.